requirements.txt
openai>=1.35.14,<2.0.0
python-dotenv
numpy
scikit-learn
matplotlib
seaborn
plotly
chromadb
tiktoken

# ==== Added for Hybrid Elasticsearch + FAISS ====
faiss-cpu
elasticsearch>=8.15.0,<9.0.0
requests
